Progress on adding global state and searchResults
