+++ frontend/src/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useContext }  from 'react';
 import './App.css';
 import { BaseProvider, DarkTheme, LightTheme, useStyletron} from 'baseui';
 import { Input, SIZE  } from "baseui/input";
@@ -6,53 +6,28 @@ import {Button} from 'baseui/button';
 import {Provider as StyletronProvider} from 'styletron-react';
 import {Client as Styletron} from 'styletron-engine-atomic';
 import axios from 'axios';
+import SearchBar from './SearchBar';
+import SearchResults from './SearchResults';
 
 const engine = new Styletron();
-
-
+const AppContext = React.createContext({});
 
 function App() {
-  const [value, setValue] = React.useState("What would you like to know?");
-  const [_, theme] = useStyletron();
-
-   async function makeSearchRequest(user_input) {
-    const response = await axios.post(
-      'http://127.0.0.1:8000/testJSON',
-      { search_query: user_input },
-      { headers: { 'Content-Type': 'application/json' } }
-    )
-    console.log(response.data)
+  const [searchResults, setSearchResults] = useState([{"url": "testing", "title": "none"}]);
+  const [resultsFound, setResultsFound] = useState(false);
+  const store = {
+    searchResults: {get: searchResults, set: setSearchResults},
+    resultsFound: {get: resultsFound, set: setResultsFound}
   }
-
+  
 
   return (
     <StyletronProvider value={engine}>
-    <BaseProvider theme={LightTheme}>
-    <div className="homepage-body">
-      <div className="homepage-search-bar">
-        <Input
-        theme={DarkTheme}
-        className="search-bar"
-        value={value}
-        onChange={e => setValue(e.target.value)}
-        size={SIZE.default}
-        placeholder=""
-        clearOnEscape
-        overrides={{
-          Root: {
-            style: {
-              width: '100%',
-              marginRight: theme.sizing.scale300,
-            },
-          },
-        }}
-      />
-      <Button
-        onClick={async() => await makeSearchRequest(value)}
-      >Search</Button>
-      </div>
-    </div>
-    </BaseProvider>
+      <BaseProvider theme={LightTheme}>
+        <AppContext.Provider value={store}>
+        {resultsFound ? <SearchResults stateContext={store}></SearchResults> : <SearchBar stateContext={store}></SearchBar>}
+        </AppContext.Provider>
+      </BaseProvider>
     </StyletronProvider>
   );
 }
