+++ frontend/src/SearchResults.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function SearchResults() {
+  return (
+  <div></div>
+  );
+}
+
+export default SearchResults;
